// src/interfaces/CountryInterface.ts
export default interface CountryInterface {
    name: Name
    tld?: string[]
    cca2: string
    ccn3?: string
    cca3: string
    cioc?: string
    independent?: boolean
    status: string
    unMember: boolean
    currencies?: Currencies
    idd: Idd
    capital?: string[]
    altSpellings: string[]
    region: string
    subregion?: string
    languages?: Languages
    translations: Translations
    latlng: number[]
    landlocked: boolean
    area: number
    demonyms?: Demonyms
    flag: string
    maps: Maps
    population: number
    fifa?: string
    car: Car
    timezones: string[]
    continents: string[]
    flags: Flags
    coatOfArms: CoatOfArms
    startOfWeek: string
    capitalInfo: CapitalInfo
    postalCode?: PostalCode
    borders?: string[]
    gini?: Gini
}

export interface Name {
    common: string
    official: string
    nativeName?: NativeName
}

export interface NativeName {
    dzo?: Dzo
    spa?: Spa
    eng?: Eng
    smo?: Smo
    fra?: Fra
    mon?: Mon
    rus?: Rus
    uzb?: Uzb
    hat?: Hat
    kon?: Kon
    lin?: Lin
    zho?: Zho
    fas?: Fas
    bis?: Bis
    est?: Est
    aym?: Aym
    que?: Que
    ara?: Ara
    run?: Run
    pau?: Pau
    cha?: Cha
    ben?: Ben
    rar?: Rar
    nep?: Nep
    tur?: Tur
    jam?: Jam
    cat?: Cat
    nrf?: Nrf
    mlt?: Mlt
    grn?: Grn
    mfe?: Mfe
    deu?: Deu
    ltz?: Ltz
    sqi?: Sqi
    srp?: Srp
    mlg?: Mlg
    urd?: Urd
    nya?: Nya
    slv?: Slv
    ssw?: Ssw
    gil?: Gil
    aze?: Aze
    swa?: Swa
    kaz?: Kaz
    lav?: Lav
    bjz?: Bjz
    ron?: Ron
    mri?: Mri
    nzs?: Nzs
    por?: Por
    sag?: Sag
    mah?: Mah
    msa?: Msa
    ber?: Ber
    fij?: Fij
    hif?: Hif
    lit?: Lit
    nld?: Nld
    pap?: Pap
    gle?: Gle
    tuk?: Tuk
    mya?: Mya
    tam?: Tam
    zdj?: Zdj
    mey?: Mey
    afr?: Afr
    her?: Her
    hgm?: Hgm
    kwn?: Kwn
    loz?: Loz
    ndo?: Ndo
    tsn?: Tsn
    tet?: Tet
    bos?: Bos
    hrv?: Hrv
    fil?: Fil
    bwg?: Bwg
    kck?: Kck
    khi?: Khi
    ndc?: Ndc
    nde?: Nde
    sna?: Sna
    sot?: Sot
    toi?: Toi
    tso?: Tso
    ven?: Ven
    xho?: Xho
    zib?: Zib
    glv?: Glv
    crs?: Crs
    ita?: Ita
    bul?: Bul
    isl?: Isl
    lua?: Lua
    ind?: Ind
    slk?: Slk
    swe?: Swe
    lao?: Lao
    heb?: Heb
    nor?: Nor
    cnr?: Cnr
    som?: Som
    sin?: Sin
    dan?: Dan
    fao?: Fao
    hmo?: Hmo
    tpi?: Tpi
    tir?: Tir
    kor?: Kor
    niu?: Niu
    khm?: Khm
    bel?: Bel
    nfr?: Nfr
    ell?: Ell
    ukr?: Ukr
    kal?: Kal
    mkd?: Mkd
    ces?: Ces
    bar?: Bar
    nau?: Nau
    div?: Div
    kat?: Kat
    pov?: Pov
    gsw?: Gsw
    roh?: Roh
    fin?: Fin
    lat?: Lat
    arc?: Arc
    ckb?: Ckb
    tkl?: Tkl
    ton?: Ton
    cal?: Cal
    hin?: Hin
    pih?: Pih
    amh?: Amh
    jpn?: Jpn
    hun?: Hun
    kir?: Kir
    vie?: Vie
    nno?: Nno
    nob?: Nob
    smi?: Smi
    kin?: Kin
    tha?: Tha
    tvl?: Tvl
    tgk?: Tgk
    nbl?: Nbl
    nso?: Nso
    zul?: Zul
    hye?: Hye
    prs?: Prs
    pus?: Pus
    pol?: Pol
}
// ...
export interface Currencies {
    [key: string]: Currency
}

export interface Currency{
    name: string,
    symbol: string
}

export interface Idd {
    root?: string
    suffixes?: string[]
}

export interface Languages {
    dzo?: string
    spa?: string
    eng?: string
    smo?: string
    fra?: string
    mon?: string
    rus?: string
    uzb?: string
    hat?: string
    kon?: string
    lin?: string
    zho?: string
    fas?: string
    bis?: string
    est?: string
    aym?: string
    que?: string
    ara?: string
    run?: string
    pau?: string
    cha?: string
    ben?: string
    rar?: string
    nep?: string
    tur?: string
    jam?: string
    cat?: string
    nrf?: string
    mlt?: string
    grn?: string
    mfe?: string
    deu?: string
    ltz?: string
    sqi?: string
    srp?: string
    mlg?: string
    urd?: string
    nya?: string
    slv?: string
    ssw?: string
    gil?: string
    aze?: string
    swa?: string
    kaz?: string
    lav?: string
    bjz?: string
    ron?: string
    mri?: string
    nzs?: string
    por?: string
    sag?: string
    mah?: string
    msa?: string
    ber?: string
    fij?: string
    hif?: string
    lit?: string
    nld?: string
    pap?: string
    gle?: string
    tuk?: string
    mya?: string
    tam?: string
    zdj?: string
    mey?: string
    afr?: string
    her?: string
    hgm?: string
    kwn?: string
    loz?: string
    ndo?: string
    tsn?: string
    tet?: string
    bos?: string
    hrv?: string
    fil?: string
    bwg?: string
    kck?: string
    khi?: string
    ndc?: string
    nde?: string
    sna?: string
    sot?: string
    toi?: string
    tso?: string
    ven?: string
    xho?: string
    zib?: string
    glv?: string
    crs?: string
    ita?: string
    bul?: string
    isl?: string
    lua?: string
    ind?: string
    slk?: string
    swe?: string
    lao?: string
    heb?: string
    nor?: string
    cnr?: string
    som?: string
    sin?: string
    dan?: string
    fao?: string
    hmo?: string
    tpi?: string
    tir?: string
    kor?: string
    niu?: string
    khm?: string
    bel?: string
    nfr?: string
    ell?: string
    ukr?: string
    kal?: string
    mkd?: string
    ces?: string
    bar?: string
    nau?: string
    div?: string
    kat?: string
    pov?: string
    gsw?: string
    roh?: string
    fin?: string
    lat?: string
    arc?: string
    ckb?: string
    tkl?: string
    ton?: string
    cal?: string
    hin?: string
    pih?: string
    amh?: string
    jpn?: string
    hun?: string
    kir?: string
    vie?: string
    nno?: string
    nob?: string
    smi?: string
    kin?: string
    tha?: string
    tvl?: string
    tgk?: string
    nbl?: string
    nso?: string
    zul?: string
    hye?: string
    prs?: string
    pus?: string
    pol?: string
}

export interface Translations {
    ara: Ara2
    ces: Ces2
    cym: Cym
    deu: Deu2
    est: Est2
    fin: Fin2
    fra: Fra2
    hrv?: Hrv2
    hun: Hun2
    ita: Ita2
    jpn?: Jpn2
    kor: Kor2
    nld: Nld2
    per?: Per
    pol: Pol2
    por: Por2
    rus: Rus2
    slk: Slk2
    spa: Spa2
    swe: Swe2
    urd: Urd2
    zho?: Zho2
}

export interface Ara2 {
    official: string
    common: string
}

export interface Ces2 {
    official: string
    common: string
}

export interface Cym {
    official: string
    common: string
}

export interface Deu2 {
    official: string
    common: string
}

export interface Est2 {
    official: string
    common: string
}

export interface Fin2 {
    official: string
    common: string
}

export interface Fra2 {
    official: string
    common: string
}

export interface Hrv2 {
    official: string
    common: string
}

export interface Hun2 {
    official: string
    common: string
}

export interface Ita2 {
    official: string
    common: string
}

export interface Jpn2 {
    official: string
    common: string
}

export interface Kor2 {
    official: string
    common: string
}

export interface Nld2 {
    official: string
    common: string
}

export interface Per {
    official: string
    common: string
}

export interface Pol2 {
    official: string
    common: string
}

export interface Por2 {
    official: string
    common: string
}

export interface Rus2 {
    official: string
    common: string
}

export interface Slk2 {
    official: string
    common: string
}

export interface Spa2 {
    official: string
    common: string
}

export interface Swe2 {
    official: string
    common: string
}

export interface Urd2 {
    official: string
    common: string
}

export interface Zho2 {
    official: string
    common: string
}

export interface Demonyms {
    eng: Eng2
    fra?: Fra3
}

export interface Eng2 {
    f: string
    m: string
}

export interface Fra3 {
    f: string
    m: string
}

export interface Maps {
    googleMaps: string
    openStreetMaps: string
}

export interface Car {
    signs?: string[]
    side: string
}

export interface Flags {
    png: string
    svg: string
}

export interface CoatOfArms {
    png?: string
    svg?: string
}

export interface CapitalInfo {
    latlng?: number[]
}

export interface PostalCode {
    format: string
    regex?: string
}

export interface Gini {
    "2017"?: number
    "2014"?: number
    "2013"?: number
    "2018"?: number
    "2019"?: number
    "2003"?: number
    "2012"?: number
    "1992"?: number
    "2011"?: number
    "2016"?: number
    "2010"?: number
    "2015"?: number
    "2009"?: number
    "1999"?: number
    "1998"?: number
    "2006"?: number
    "2008"?: number
    "2005"?: number
    "2004"?: number
}